Update README.md
The Router Module
It's a wrapper component for all the components that will be using client-side routing to render themselves, and it informs React about that. <Switch/> ensures that only one route is handled at a time, and the <Route/> component tells us which component will be rendered on which route

This Demo Project includes following Pages:- Login.js Form ,Signup.js Form and Dashbord.js Page.